dev-requirements.txt
boto==2.43.0
Django==1.10.2
django-storages-redux==1.3.2
Pillow==3.4.2
mysqlclient==1.3.7
django-haystack
Whoosh
